Add global cycle pos cmds and show cycle in GUI
This commit adds the commands "global_cycle_len" and "global_cycle_pos"
to sooperlooper. The current cycle position is displayed as a bar in the
quantize GUI element. This feature only works with Internal sync at the
moment, as this is the only case where sooperlooper has an internal
cycle pos (JackSync has a cycle length, but no position). Quantize has
to be set to cycle or 8th. The current cycle is only displayed when it
is >= 1s to avoid having too much visual noise in the GUI.

When recording the first loop in sync internal + quantize cycle + first
loop on sync it is hard to hit the first note at the beginning of the
cycle. A visual indicator of where we are in the cycle helps to be on
time in situations where no other cue is available.

ChoiceBox has been augmented with the feature to draw a bar in its
background. It is similar to what SliderBar can do but has less options
(e.g. it only displays values from 0.0f to 1.0f, no value text
rendering).