Add setting to copy loop flags on loop create
A new setting now controls if a newly created loop inherits all flags
from the previous loop (sync, play sync, feedback play, tempo stretch,
discrete prefader). This is a gui-only implementation. The MainPanel
class has an extra counter to track when this instance of slgui has
created a panel, to then copy the values over. This was done to not
interfere with other sooperlooper clients that might not wish the values
to be copied.